rename modal now flag is removed
