+++ frontend/src/lib/components/AddToDashboard/AddToDashboard.scss
